+++ CHANGELOG.md
@@ -9,7 +9,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Changed componenet `flexberry-darepicker` classNames: add class `flexberry-datepicker`.
 * Component's template:
   * Changed readonly attribute for `flexberry-darepicker`.
-  * Changed button's classes.
+  * Changed button's classes `flexberry-darepicker`.
+  * Changed button's classes `groupedit-toolbar`.
 
 ## [0.9.0-beta.9] - 2017-05-19
 ### Added